Add component test for stats and fix a lotta bugs

+++ core/src/main/java/org/nzbhydra/web/mapping/stats/IndexerApiAccessStatsEntry.java
@@ -7,6 +7,6 @@ public class IndexerApiAccessStatsEntry {
     private String indexerName = null;
     private Double percentSuccessful = null;
     private Double percentConnectionError = null;
-    private Integer averageAccessesPerDay = null;
+    private Double averageAccessesPerDay = null;
 
 }
